// Frontend/src/types/generated/models/Address.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type Address = {
    readonly id: string;
    readonly createdAt: string;
    readonly updatedAt: string;
    street: string;
    city: string;
    stateOrProvince: string | null;
    postalCode: string;
    country: string;
    latitude: number | null;
    longitude: number | null;
};