// Frontend/src/types/generated/models/Branch.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { Address } from './Address';

export type Branch = {
    id: string;
    address: Address;
    readonly createdAt: string;
    readonly updatedAt: string;
    name: string;
    phone: string;
    email: string | null;
    company: string;
};